backend/requirements.txt
aiortc==1.10.1
fastapi==0.115.8
mediapipe==0.10.9
numpy==1.23.5
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
pymongo==4.11.1
python-dotenv==1.0.1
torch==2.6.0
torchvision==0.21.0
